requirements.txt
streamlit
pandas
scikit-learn
joblib
matplotlib
seaborn
plotly
langchain
langchain-core
langchain-community
langgraph
langchain-huggingface
sentence-transformers
faiss-cpu
python-dotenv
langchain-openai
